rename/move internal headers
