+++ Poisson_surface_reconstruction_3/include/CGAL/Surface_mesher/Poisson_implicit_surface_oracle_3.h
@@ -18,7 +18,7 @@
 
 
 #include <CGAL/point_generators_3.h>
-#include <CGAL/Surface_mesher/Poisson_sphere_oracle_3.h>
+#include <CGAL/Poisson_surface_reconstruction_3/internal/Poisson_sphere_oracle_3.h>
 #include <CGAL/Surface_mesher/Implicit_surface_oracle_3.h>
 #include <CGAL/Real_embeddable_traits.h>
 #include <CGAL/squared_distance_3.h>
